refactor: remove aggregators from global_ranking
