+++ quanda/explainers/__init__.py
@@ -3,12 +3,6 @@
 from quanda.explainers import utils, wrappers
 from quanda.explainers.base import Explainer
 from quanda.explainers.functional import ExplainFunc, ExplainFuncMini
-from quanda.explainers.global_ranking.aggregators import (
-    AbsSumAggregator,
-    BaseAggregator,
-    SumAggregator,
-    aggr_types,
-)
 from quanda.explainers.random import RandomExplainer
 
 __all__ = [
@@ -18,8 +12,5 @@ __all__ = [
     "ExplainFuncMini",
     "utils",
     "wrappers",
-    "BaseAggregator",
-    "SumAggregator",
-    "AbsSumAggregator",
-    "aggr_types",
+    "global_ranking",
 ]
